+++ elements/map/src/methods/map/index.js
@@ -8,6 +8,7 @@ export {
   setConfigMethod,
   setProjectionMethod,
   setSyncMethod,
+  setTerrainMethod,
 } from "./setters";
 export {
   getLonLatCenterMethod,