+++ elements/map/src/plugins/globe/openglobus.js
@@ -8,6 +8,8 @@ import {
   OpenStreetMap,
   Vector,
   Entity,
+  GlobusRgbTerrain,
+  EmptyTerrain,
 } from "@openglobus/og";
 import { buildExpression, newEvaluationContext } from "ol/expr/cpu.js";
 import { distributeTileToIdealMap } from "./methods.js";
@@ -79,12 +81,20 @@ export const createGlobe = ({ EOxMap, target, mapPool }) => {
     sun: { active: false },
     atmosphereEnabled: false,
     transparentBackground: true,
+    terrain: EOxMap.globeConfig?.terrain
+      ? new GlobusRgbTerrain()
+      : new EmptyTerrain(), // Use EOxMap's globeConfig.terrain property
     resourcesSrc: "https://cdn.jsdelivr.net/npm/@openglobus/og@0.27.21/lib/res",
     fontsSrc:
       "https://cdn.jsdelivr.net/npm/@openglobus/og@0.27.21/lib/res/fonts",
   });
+  let currZoom = EOxMap.map.getView().getZoom() - 1;
+  if (!EOxMap.globeConfig?.useHighLOD) {
+    globus.planet.quadTreeStrategy.setLodSize(512);
+    currZoom = currZoom - 1;
+  }
 
-  const height = 21050000 / Math.pow(2, EOxMap.map.getView().getZoom() - 1);
+  const height = 21050000 / Math.pow(2, currZoom);
   const center = EOxMap.map.getView().getCenter();
   const newCenter = transform(
     center,
@@ -410,8 +420,9 @@ export const disableGlobe = (map) => {
         const finalCameraPosition = globe.planet.camera.getLonLat();
 
         // Calculate the OpenLayers zoom level using the camera's final height
+        let zoomFactor = map.globeConfig?.useHighLOD ? 1 : 2;
         const zoomFromGlobe =
-          Math.log2(21050000 / finalCameraPosition.height) + 1;
+          Math.log2(21050000 / finalCameraPosition.height) + zoomFactor;
 
         // Calculate the OpenLayers center coordinates
         const centerFromGlobe = [