+++ elements/map/stories/map.stories.js
@@ -54,6 +54,7 @@ export default {
       .layers=${args.layers}
       .projection=${args.projection}
       .zoom=${args.zoom}
+      .globeConfig=${args.globeConfig}
     ></eox-map>
   `,
 };
@@ -75,7 +76,7 @@ export const Primary = PrimaryStory;
  * Instead of passing each property individually, one can also pass a `config` property:
  * Note two things here: when using the `config` property, then `center` and `zoom` are nested inside the `view` property. Secondly, this example includes `controls`: by passing `Zoom: {}` this enables the OpenLayers `Zoom` control, without any options (which could be passed inside the object).
 * `config` supports the following properties:
-* 
+*
 * ```ts
 * export type ConfigObject = {
 *   controls: ControlDictionary;
@@ -92,7 +93,7 @@ export const Primary = PrimaryStory;
 *   animationOptions?: EOxAnimationOptions;
 * };
 * ```
-* 
+*
 * ```ts
 * export type ControlDictionary = {
 *   Zoom?: ConstructorParameters<typeof import("ol/control/Zoom").default>[0];
@@ -365,6 +366,7 @@ export const HelperMethods = HelperMethodsStory;
 /**
  * Basic Globe rendered using `projection: "globe"`, and OpenGlobus as globe renderer.
  * When the projection of the view can is set to "globe", the map will render as a 3D globe.
+ * The `globeConfig` property allows to configure the globe, e.g. setting `terrain` to `true` to enable terrain rendering, or `useHighLOD` to `true` for higher level of detail.
  * This example renders a globe centered on Austria, with 2 layers: an XYZ tile layer as base layer, and a GeoTiff layer on top.
  * The GeoTiff layer is rendered as a CanvasTiles layer in OpenGlobus, while the XYZ layer is rendered natively in the OpenGlobus.
  */